MDL-11521 fixed problems when creating contexts and aprents does not exist yet MDL-11527 fixed warning from $CFG->defaultuserroleid when installing and upgrading + removed reusing of previous contexts in has_capability() and get_context_instance()  because it was hiding critical errors when context was false, not specified, etc. + tweaked roles defaults - faster and defaults are applied when roles exist + SYSCONTEXTID may be specified in config.php to eliminate 1 db query + static context cache is now reset after rebuilding of context paths + path field in context table is now nullable - pg compatibility requirement + other minor fixes

+++ lib/db/mysql.php
@@ -2195,7 +2195,7 @@ function main_upgrade($oldversion=0) {
                     )TYPE=MYISAM COMMENT ='time user last accessed any page in a course';", true);
     }
 
-    if ($oldversion < 2006091212) {   // Reload the guest roles completely with new defaults
+    if (!empty($CFG->rolesactive) and $oldversion < 2006091212) {   // Reload the guest roles completely with new defaults
         if ($guestroles = get_roles_with_capability('moodle/legacy:guest', CAP_ALLOW)) {
             delete_records('capabilities');
             $sitecontext = get_context_instance(CONTEXT_SYSTEM, SITEID);
@@ -2218,7 +2218,7 @@ function main_upgrade($oldversion=0) {
         delete_records('config', 'name', 'requestedstudentsname');
     }
 
-    if ($oldversion < 2006091901) {
+    if (!empty($CFG->rolesactive) and $oldversion < 2006091901) {
         if ($roles = get_records('role')) {
             $first = array_shift($roles);
             if (!empty($first->shortname)) {
